+++ Sources/KeyboardKit/Views/Input/KeyboardTextView.swift
@@ -1,5 +1,5 @@
 //
-//  KeyboardTextField.swift
+//  KeyboardTextView.swift
 //  KeybaordKit
 //
 //  Created by Daniel Saidi on 2021-07-15.
@@ -11,22 +11,25 @@ import UIKit
 
 /**
  This view can be used within a keyboard extension, when you
- want to provide a single-line text field next to a keyboard
+ want to provide a multi-line text view next to the keyboard
  and have that receive user input instead of the hosting app.
  
  The view will automatically register itself as an alternate
  proxy when it becomes first responder and unregister itself
  when it resigns first responder. This makes it receive user
  actions instead of the hosting app.
+ 
+ `NOTE` that you must give this view a specific `height` for
+ it to show up, otherwise it will collapse to zero height.
  */
-public struct KeyboardTextField: UIViewRepresentable {
+public struct KeyboardTextView: UIViewRepresentable {
     
     /**
-     Create a keyboard text field instance.
+     Create a keyboard text view instance.
      
      - Parameters:
        - resignOnReturn: Whether or not to resign first responder when return is pressed.
-       - config: A configuration block that can be used to configure the text field.
+       - config: A configuration block that can be used to configure the text View.
      */
     public init(
         resignOnReturn: Bool = true,
@@ -38,37 +41,35 @@ public struct KeyboardTextField: UIViewRepresentable {
     private let config: ConfigAction
     private let resignOnReturn: Bool
     
-    public typealias ConfigAction = (UITextField) -> Void
+    public typealias ConfigAction = (UITextView) -> Void
     
-    public func makeUIView(context: Context) -> UITextField {
-        let textField = KeyboardInputTextField()
-        textField.resignOnReturn = resignOnReturn
-        config(textField)
-        return textField
+    public func makeUIView(context: Context) -> UITextView {
+        let textView = KeyboardInputTextView()
+        textView.resignOnReturn = resignOnReturn
+        config(textView)
+        return textView
     }
     
-    public func updateUIView(_ uiView: UITextField, context: Context) {}
+    public func updateUIView(_ uiView: UITextView, context: Context) {}
 }
 
-class KeyboardInputTextField: UITextField {
+class KeyboardInputTextView: UITextView, KeyboardInputTextComponent {
     
     var resignOnReturn: Bool = true
     
     override func insertText(_ text: String) {
-        if resignOnReturn && text == "\n" {
-            _ = resignFirstResponder()
-            return
-        }
+        guard handleInsertText(text) else { return }
         super.insertText(text)
     }
     
     override func becomeFirstResponder() -> Bool {
-        KeyboardInputViewController.shared.textInputProxy = TextInputProxy(input: self)
+        handleBecomeFirstResponder()
         return super.becomeFirstResponder()
     }
     
+    @discardableResult
     override func resignFirstResponder() -> Bool {
-        KeyboardInputViewController.shared.textInputProxy = nil
+        handleResignFirstResponder()
         return super.resignFirstResponder()
     }
 }